Removed compiler warning for Xcode 4.5 64-bit

+++ cocos2d/CCAction.m
@@ -62,7 +62,7 @@
 
 -(NSString*) description
 {
-	return [NSString stringWithFormat:@"<%@ = %p | Tag = %i>", [self class], self, tag_];
+	return [NSString stringWithFormat:@"<%@ = %p | Tag = %ld>", [self class], self, (long)tag_];
 }
 
 -(id) copyWithZone: (NSZone*) zone
